# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Snyder Dialer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{id}/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /campaigns/{id}/leads/upload:
    post:
      operationId: importLeadsCsv
      tags: [leads]
      summary: Upload leads CSV for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/LeadsCsvUpload"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
  /leads/{id}:
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: No content

  # Knowledge Base
...
  /prompt-templates:
    get:
      operationId: listPromptTemplates
      tags: [prompt-templates]
      summary: List prompt templates
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptTemplate"
    post:
      operationId: createPromptTemplate
      tags: [prompt-templates]
      summary: Create a prompt template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptTemplate"
  /prompt-templates/{id}:
    patch:
      operationId: updatePromptTemplate
      tags: [prompt-templates]
      summary: Update a prompt template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptTemplateUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptTemplate"
    delete:
      operationId: deletePromptTemplate
      tags: [prompt-templates]
      summary: Delete a prompt template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: No content
components:
  schemas:
...
    Lead:
      type: object
      required: [id, campaignId, name, phone, status, createdAt]
      properties:
        id:
          type: number
        campaignId:
          type: number
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, calling, completed, failed, do_not_call]
        createdAt:
          type: string
          format: date-time
    LeadsCsvUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
    UploadResult:
      type: object
      required: [imported, skipped, total]
      properties:
        imported:
          type: number
        skipped:
          type: number
        total:
          type: number
...
    PromptTemplate:
      type: object
      required: [id, name, content, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        content:
          type: string
        variables:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    PromptTemplateInput:
      type: object
      required: [name, content]
      properties:
        name:
          type: string
          minLength: 1
        content:
          type: string
        variables:
          type: array
          items:
            type: string
    PromptTemplateUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        content:
          type: string
        variables:
          type: array
          items:
            type: string