# lib/api-zod/src/generated/types/promptTemplate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */

export interface PromptTemplate {
  id: number;
  name: string;
  content: string;
  variables?: string[];
  createdAt: Date;
}