# lib/api-zod/src/generated/types/promptTemplateUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */

export interface PromptTemplateUpdate {
  /** @minLength 1 */
  name?: string;
  content?: string;
  variables?: string[];
}