# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Snyder Dialer API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  imported: number;
  skipped: number;
  total: number;
}